Refactor code to have Styled js files
